feat: auth 관련 디렉토리 정리 및 provider 배치
