+++ src/pages/_app.page.tsx
@@ -7,6 +7,7 @@ import { Hydrate, QueryClient, QueryClientProvider } from '@tanstack/react-query
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { domMax, LazyMotion } from 'framer-motion';
 
+import AuthProvider from '~/components/auth/AuthProvider';
 import ErrorBoundary from '~/components/error/ErrorBoundary';
 import MonitoringInitializer from '~/components/monitoring/MonitoringInitializer';
 import ToastWrapper from '~/components/toast/ToastWrapper';
@@ -45,10 +46,12 @@ export default function App({ Component, pageProps }: AppPropsWithLayout) {
             <LazyMotion features={domMax}>
               <GlobalStyles />
               <ErrorBoundary>
-                <div css={defaultLayoutCss}>
-                  {getLayout(<Component {...pageProps} />)}
-                  <ToastWrapper />
-                </div>
+                <AuthProvider>
+                  <div css={defaultLayoutCss}>
+                    {getLayout(<Component {...pageProps} />)}
+                    <ToastWrapper />
+                  </div>
+                </AuthProvider>
               </ErrorBoundary>
             </LazyMotion>
           </ThemeProvider>
